Store loaded Config::Any to class accessor
This refactor makes publicly available and documents the config class attribute that
is loaded by DBIC::Schema::Config when searching for connection info. this allows
DBIC::Schema::Config to be used as a more general purpose config loader for DBIC -
not just for connection information.

I've left the config lazyloaded until connection() is called - if $schema->config
is called before that it will cache and load for later.