Improve caching layer
Previously config() would return a reference to whatever was loaded from
_load_config().  Changes to that structure would result in changes to the
return value of ->config leading to an invalid cache issue.

The dclone ensures that config will persist in returning the value from
_load_config().

Storable has been core since 5.7.3, and a result has not been added to
the Makefile.PL as a requirement.